fix: change order or complete onboarding action and move schema to schemas folder
