+++ web-app/src/lib/schemas/index.ts
@@ -3,4 +3,5 @@ export * from "./auth";
 export * from "./credit";
 export * from "./invitation";
 export * from "./job";
+export * from "./onboarding";
 export * from "./organization";
